fix: NavigationのExampleに余計なdivが含まれていたので削除
